pyproject.toml
[project]
name = "physics760-computational-physics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "scipy>=1.16.2",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
]
